Feature/work

+++ src/main/java/com/oven/server/api/user/domain/RatingWork.java
@@ -23,9 +23,9 @@ public class RatingWork extends BaseEntity {
     @JoinColumn(name = "work_id")
     private Work work;
 
-    private int rating;     // 1~5점
+    private Float rating;     // 1~5점
 
-    public RatingWork(User user, Work work, int rating) {
+    public RatingWork(User user, Work work, Float rating) {
         this.user = user;
         this.work = work;
         this.rating = rating;
